WorkerConsoleLogService shouldn't use ILoggerFactory
The ILoggerFactory now contains OpenTelemetry loggers, which will results
in messages being re-published to OpenTelemetry when they are meant just
for console in this Service.
Instead, take a ConsoleLoggerProvider as input and if one isn't provided
(Console Logging is disabled) then use a NullLogger. Use this Provider
directly, meaning it logs *only* to the Console (and no other registered
loggers)