Removing parsing of host.json into otel configuration
we're going to rely only on officially-supported schema for configuring otel
and as of today there isn't one for configuration files.

So, this will configure otel entirely via environment variables, natively.
The hope is in the future this code will be able to leverage config file loading
via the Otel.NET sdk vs rolling our own, but some massaging may be necessary
(like we're doing for the traces/metrics configurations) depending on how
things shake out.